Run gofmt
