+++ cmd/tgptbot/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "telegram-chatgpt-bot/internal/bot"
+
+func main() {
+	bot.Run()
+}
