--- main.go
@@ -1,6 +1,0 @@
-// main.go
-package main
-
-func main() {
-    runBot()
-}
